refactor: improve budget module structure, pagination, and code quality

+++ src/main/java/com/example/intelliwealth/fynix/application/dto/FinancialSummary.java
@@ -14,7 +14,7 @@ public class FinancialSummary {
     private int activeSubscriptions;
     private double monthlySubscriptionCost;
 
-    private int totalGoals;
+    private long totalGoals;
     private double goalProgress; // Average %
 
     // New field to carry context to the AI
